[test] 로그인 유지 오류 해결 배포 환경 테스트
