+++ frontend/src/components/common/ProtectedRoutes/index.tsx
@@ -1,10 +1,8 @@
-import React, { useEffect } from 'react';
-import { unwrapResult } from '@reduxjs/toolkit';
-import { Navigate, Outlet, useNavigate } from 'react-router-dom';
+import React from 'react';
 
-import { loginCheck } from '@redux/modules/authSlice/thunk';
-import { useAppDispatch, useAppSelector } from '@hooks/reduxHooks';
-import { StorageType, setValue } from '@utils/storageUtils';
+import { Navigate, Outlet } from 'react-router-dom';
+
+import { useAppSelector } from '@hooks/reduxHooks';
 
 import LoadingPage from '@pages/LoadingPage';
 
@@ -21,33 +19,8 @@ const ProtectedRoutes = ({
 }: ProtectedRoutesProps) => {
   const { isLogin, loading } = useAppSelector(state => state.auth);
   const { role } = useAppSelector(state => state.user);
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
 
-  useEffect(() => {
-    const redirectToMainPage = () => {
-      navigate('/');
-    };
-
-    const fetchLoginCheck = async () => {
-      try {
-        const resultAction = await dispatch(loginCheck());
-        const data = unwrapResult(resultAction);
-        if (data.userInfo.doNotLogout) setValue(StorageType.LOCAL, 'userInfo', data.userInfo);
-        else setValue(StorageType.SESSION, 'userInfo', data.userInfo);
-      } catch (error) {
-        redirectToMainPage();
-      }
-    };
-
-    if (document.cookie.split('; ').find(row => row.startsWith('access_token='))) {
-      fetchLoginCheck();
-    }
-  }, [dispatch, isLogin]);
-
-  if (loading) {
-    return <LoadingPage />;
-  }
+  if (loading) return <LoadingPage />;
 
   if (isLogin && blockLogin) return <Navigate to="/" />;
   if (requireAuth && !isLogin) return <Navigate to="/login" />;
